Use delayload to demo missing library feature
Add `xrt::version()` API. Pretend the latest version of xrt library is
`2` and that this library has all the latest APIs.

Add `xrt_legacy` directory that removes `xrt::run()` API.  Pretend
this is version `1` of the xrt library.

Condition `vart.cpp` to call `xrt::run()` only if `xrt::version() >=2`.

Delayload `xrt.dll`.  Default is version 2 of the library, which makes
vart call `xrt::run()`.  Overwrite xrt.dll with the legacy version and
see that vart no longer calls `xrt::run()`.

Point is that delay load can be used with an application that is
compiled against latest version of a library. Same application
binary can be used with an older deployed version of the library.